refactor(engine): 外移 pi.rs 内联测试至 pi_tests.rs
- 4318→3201 行，纯搬运零行为变更
- baseline 重生成：pi.rs reduced -1117；新文件登记豁免
验证：cargo test --lib engine::pi 98 passed；cargo check 通过；gate 绿
